+++ packages/core/src/BreadCrumb/styles.js
@@ -25,6 +25,9 @@ const styles = theme => ({
   separatorContainer: {
     width: "32px",
     height: "32px"
+  },
+  orderedList: {
+    display: "flex"
   }
 });