+++ packages/core/src/theme/palette.js
@@ -15,7 +15,7 @@ const palette = theme =>
       disabled: theme.palette.atmosphere.atmo7
     },
     background: {
-      default: theme.palette.atmosphere.atmo2
+      default: theme.palette.atmosphere.atmo3
     },
     ...Object.assign({}, ...Object.values(theme.palette))
   });