+++ packages/doc/samples/templates/login/styles.js
@@ -1,7 +1,6 @@
 const styles = theme => ({
   root: {
-    height: "100vh",
-    backgroundColor: theme.hv.palette.atmosphere.atmo2
+    height: "100vh"
   }
 });