+++ automation/storybook/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 # [2.0.0-next.0](https://github.com/pentaho/hv-uikit-react/compare/@hv/uikit-react-automation-storybook@1.8.8...@hv/uikit-react-automation-storybook@2.0.0-next.0) (2020-02-14)
+# [1.10.0](https://github.com/pentaho/hv-uikit-react/compare/@hv/uikit-react-automation-storybook@1.9.0...@hv/uikit-react-automation-storybook@1.10.0) (2020-02-28)
+
+
+### Features
+
+* **tabs:** implement the component according to 1.2.1 DS version. [#245](https://github.com/pentaho/hv-uikit-react/issues/245) ([d7d5653](https://github.com/pentaho/hv-uikit-react/commit/d7d5653a52e230fa2f0d871468fa4c06947e42c9))
+
+
+
+
+
 # [1.9.0](https://github.com/pentaho/hv-uikit-react/compare/@hv/uikit-react-automation-storybook@1.8.9...@hv/uikit-react-automation-storybook@1.9.0) (2020-02-20)