+++ packages/core/CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 # [2.0.0-next.0](https://github.com/pentaho/hv-uikit-react/compare/@hv/uikit-react-core@1.23.0...@hv/uikit-react-core@2.0.0-next.0) (2020-02-14)
+# [1.25.0](https://github.com/pentaho/hv-uikit-react/compare/@hv/uikit-react-core@1.24.0...@hv/uikit-react-core@1.25.0) (2020-02-28)
+
+
+### Bug Fixes
+
+* **breadcrumb:** accessibility review. [#1082](https://github.com/pentaho/hv-uikit-react/issues/1082) ([5d94d79](https://github.com/pentaho/hv-uikit-react/commit/5d94d79951b37f12b16226c9afa2b33fd7d1acb2))
+* **card:** inject accessibility rules for actions and dropdown menu [#922](https://github.com/pentaho/hv-uikit-react/issues/922) ([78e3d30](https://github.com/pentaho/hv-uikit-react/commit/78e3d3040fc84c85f689a699966a91d53b4d7ada))
+* **fileuploader:** change icon color. [#1091](https://github.com/pentaho/hv-uikit-react/issues/1091) ([6660295](https://github.com/pentaho/hv-uikit-react/commit/6660295aa768ee92848aeb9959e8adf42abc1944))
+* **templates:** alignment and colors. [#1174](https://github.com/pentaho/hv-uikit-react/issues/1174) ([024138e](https://github.com/pentaho/hv-uikit-react/commit/024138e8c163d9d80dbc463ff448c431e3a7152e))
+* **verticalnavigation:** overwrite box style. [#1170](https://github.com/pentaho/hv-uikit-react/issues/1170) ([2b0adfc](https://github.com/pentaho/hv-uikit-react/commit/2b0adfcf4627f211a1d16084d68b91761e82acf3))
+
+
+### Features
+
+* **tabs:** implement the component according to 1.2.1 DS version. [#245](https://github.com/pentaho/hv-uikit-react/issues/245) ([d7d5653](https://github.com/pentaho/hv-uikit-react/commit/d7d5653a52e230fa2f0d871468fa4c06947e42c9))
+
+
+
+
+
 # [1.24.0](https://github.com/pentaho/hv-uikit-react/compare/@hv/uikit-react-core@1.23.1...@hv/uikit-react-core@1.24.0) (2020-02-20)